fix stream_event for SimpleAgentGenerator & ToolGenerator
